Various improvements to the configurator
