--- Assets/MRTK/Core/Inspectors/Setup/XRPipelineConfigWindow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f9e73b34228b3ee4e9e7b29417141d0a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
